Refactoring: Implementando Optional ao método DELETE da classe TopicosController, para devolver status 404 caso o cliente passe na url um id que não existe no banco de dados

+++ src/main/java/br/com/alura/forum/controller/TopicosController.java
@@ -82,8 +82,13 @@ public class TopicosController {
 	@DeleteMapping("/{id}")
 	@Transactional
 	public ResponseEntity<?> remover(@PathVariable Long id){
-		topicoRepository.deleteById(id);
-		return ResponseEntity.ok().build();
+		Optional<Topico> topico = topicoRepository.findById(id);
+		if(topico.isPresent()) {
+			topicoRepository.deleteById(id);
+			return ResponseEntity.ok().build();
+		}
+		
+		return ResponseEntity.notFound().build();
 	}
 
 }
